Refactor: ToastContext -> toastContext 이름 변경
